Delay cache file creation
When called with -C/--directory, the cache directory and file would be
created in the directory where the process was started and not in the
directory where it should be executed.

Delaying the creation of the cache file until the cache is about to be
loaded makes it be created after the process has done its chdir.

Closes #1.