Clarify scheduler behaviour
Clarify that it is valid for the scheduler to return an empty jobs list
even if there are no jobs currently running and there are more jobs left
to run if all jobs in the current ready list were skipped.

Runners should continue polling for jobs until the scheduler is no
longer in "running" state.