Environments feature
Add support for evaluating Python Byggfiles in their own respective
environments.

Environments are declared in Byggfile.yml, and each environment can
specify which virtual environment its Python Byggfile (and in turn, the
actions specified therein) should be executed in.

Each environment specifies which command to use to create the virtual
environment as well as which files to consider as inputs (e.g. the
requirements files).

Bygg will compute a digest for the inputs and recreate the environment
if the input files change.

Closes #2.