Add entrypoint completion with argcomplete
In this implementation, individual environments in Byggfile.yml will not
be loaded to find entrypoints declared in Python in each environment. In
this case, only entrypoints declared in Byggfile.yml will show up for
completion.

However, if the configuration only have actions declared in Byggfile.py,
it will be loaded to find completions from it.

This is also how the entrypoint lookup works for building.

Also: add a description field to the Action class.

Closes #37.