Remove scheduler and runner global objects
Don't create the runner and scheduler objects in their respective source
files. Instead create them explicitly from where they're going to be
used, and connect them to each other there.

Set the scheduler object as a class variable on the Action class when
the scheduler is initialised instead of using the global.

This is a cleaner and more decoupled design that improves testability of
the system.