Add functions for running subprocesses connected to a tty
The subprocess module in the Python standard library does not have a
convenient way of getting the output of a subprocess as it is produced.
Having this ability is useful both for displaying the output of a
subprocess in real time, as well as for adding metadata like timestamps
to the output line by line.

Add subprocess_tty, a function for running subprocesses connected to a
tty. The function returns a generator object that yields the output of
the subprocess line by line as it is produced. After the generator has
been exhausted, the object's returncode attribute contains the
returncode of the subprocess. With credit to
https://stackoverflow.com/a/77387332 .

Add a convenience function, subprocess_tty_print, for calling a process
and print out the output as it is produced.

Add another convenience function, call, that works like the former but
in addition returns a customised int whose boolean value is True if the
subprocess returned a zero exit code, and False otherwise. Intended for
use e.g. when using Python for shell scripting.