Only try to partition Byggfile.yml actions if there are any
Building the transform action from examples/trivial would fail with a
"Could not find action" error, but building the same action as the
default action would work.

This was due to the partitioning code trying to partition the empty list
of actions from Byggfile.yml when there was an action given on the
command line, and this would fail.

The default action case would succeed because
examples/trivial/Byggfile.yml has a default action declared.

Fix this in the partitioning code so that Byggfile.py will be evaluated.

Also add a test case to noxfile.py.