Add tests for --list and --tree
Add tests for the --list and --tree functionality using pytest, and
check the output using snapshots.

Include notes for tests that are known to fail or pass when they
shouldn't.

Add a fixture for setting up a clean copy of the repo files a temporary
directory, and run the tests there. This is needed in case they behave
differently the first and subsequent times they are run.