Cleanup and fixes to entrypoints and dispatching
Correct the behaviour of default actions, including with the --list and
--tree arguments. These would not work in all combinations of having a
default target set and actions declared in YAML or Python.

Details:

dispatcher cleanup and bug fixes:

Resolve correctly to the default action declared in Byggfile.yml when
loading the actual actions from a Python file.

Make the handling of --list and --tree clearer and make them work
correctly.

get_entrypoints:

Have get_entrypoints return entrypoints from the configuration if the
scheduler has not yet been initialised. This happens when environments
in addition to the default environment have been configured in the
configuration file. In this case, all entrypoints must currently be
declared in the configuration file (this constraint is not new from this
commit).

Also add a small abstraction to have it only return the names and
descriptions of the actions.

Other cleanup:

- Rename the SettingsSection class to Settings.
- Create a default Settings structure if there is no settings file.
- Include the Settings structure in ByggContext.
- Simplify code that can now assume that ByggContext.configuration
  always exists.
- Adapt entrypoint_completions accordingly.