Renovate shell tab completion
This commit contains a general overhaul of the shell tab completion
integration:

Subclass argcomplete's CompletionFinder and override _get_completions,
since there is a need to filter out arguments depending on which other
arguments that are already on the command line. Not sure if the same
could be achieved by just reworking the argparse usage; this should be
investigated more closely.

Add a custom directories completer that will complete directories with a
Byggfile{.py,.yml} in them.

Add completion for entrypoint actions. This is also supported when there
is subdirectory declared with -C/--directory.

Output a completions file generated by argcomplete to the directory
where the bygg main file is installed.

Add a --completions argument that instructs the user how to activate the
completions in their shell.

Add tests.

Update documentation.