Resolve preamble module load path
Resolve the module load path that is inserted into the preamble to the
Python Byggfiles. This makes the paths work reliably even when they're
symlinked.

This solves the test failures that have started occurring in the GitHub
Actions in CI, but should also be a general improvement.

Also break out the preamble to a multiline string for improved
readability.