Add test for --version
Add a simple test for --version to check that the code runs and that the
version has numbers in at least the major and minor semver position. It
is not possible to test the version fully since it is managed
dynamically by the release process and thus will vary depending on the
repository state.