Lock Python versions
Lock Python versions down to patch versions 3.11.9 and 3.12.3 for mise
and the code_quality GitHub workflow.

Have nox use versions locked to minor versions 3.11 and 3.12 in
noxfile.py so that it will use the latest matching version that is
installed.

In the code_quality workflow, run nox with the explicit Python versions
defined in the workflow.

Closes #138.