Create status directory before writing cache file
The cache directory would not have been created before saving the cache
if not building any actions in a project, causing a crash.

Discovered in the nox logs for examples/only_python/. The crash would
not cause the nox job to fail because the case where it happened
actually should exit with the return code 1.

Add a call to make_sure_status_dir_exists.

Regression introduced in 4403ed67e267067c996004448d1f1ca6fed155ff.