Mypy-related fixes
Upgrade mypy -> 1.14.0

Exclude the untyped dc_schema package from mypy checking.

Turn off the error code func-returns-value since it gives a false
positive for ProcessGenerator.__iter__ in system_helpers.py. See
https://github.com/python/mypy/issues/15785#issuecomment-1659925682 .

Change references to the argcomplete typing problem to point to the
issue in this repo, and turn off the other instance of it that mypy
found.

Update the fix for the list invariance problem that pyright found to a
solution that also passes mypy. Followup to
d32a11c1c0417f4974a3f7322449043e9dfd85fa .

Fix a problem where a variable named action was reused by mistake, and
with a different type.