Rewrite dispatching mechanism
The dispatcher code has grown organically since the beginning of the
project and has been patched along the way to include new features. This
has made it increasingly harder to maintain and add new features to.

This change is a rewrite of the dispatching mechanism. This includes the
way that environments are loaded and how actions are executed in them as
well as in the top-level and subprocess dispatch.

Also, there is now no longer a default environment. Actions declared in
the static configuration are executed in the ambient environment unless
they also have an environment declared.