Rework gathering information and running action
The flows for gathering information and running actions only differ in
the build and clean step. Unify collect_environment_data and
run_action_in_environment into the same function.

Rework the flow in the new function so that the gathering and the
running is done in the ambient environment if no restart is needed. This
case would previously abort with an error message, making it impossible
to gather or run if the correct environment was already activated in the
shell.