Store environment name in each action
Store the name of the environment it belongs to in each action so that
the command line commands will only operate on actions from the correct
environment.

Previously, actions belonging to the ambient environment would be listed
also under any other environment if that environment was already
activated in the shell.