Start with the right environment if known
When an action is declared in the static configuration it is known what
environment it should be run in. Put that environment first in the list
of environments when running the action.

Add a parameter to the test_main reset test setup so that the
environments example will trigger installing all its environments.