Force uv to use the Python from the current PATH
When running bootstrap.py, uv would install the lowest version of Python
that it could find based on the Python requirements set in
pyproject.toml.

Adding --no-project and --no-managed-python leaves uv with no choice but
to use the Python that is available in the current PATH. This makes the
process compatible with using e.g. mise for managing Python versions.