Handle failed jobs at dispatcher top level
Propagate failed jobs status to the top level of the dispatcher instead
of doing sys.exit. This makes --watch work also for failed jobs.
Followup to 29cd1c76f9af41e299eec65f052b801c791b2a78.

Also output a status summary of the failed jobs.

This commit changes the status code of --check to be what the relevant
build job returned instead of always being 1, and the relevant tests
have been changed to reflect that. This is more of a side effect of the
current implementation than it is a conscious decision on what --check
actually should return.