Use loky for multiprocess handling
The runner uses the third-party multiprocess package for managing the
process pool because it needs a package with better serialising
capabilities than pickle (it uses dill). However, multiprocess is a
quite heavy dependency, and one has to poll the job queue actively.

Use loky instead, which has futures that can be waited on. It uses
cloudpickle for serialising, and is also a much smaller dependency.

This change reduces the overhead for the job queue loop approximately by
half, even though that is hardly noticeable in scenarios with realistic
load.