Move argcomplete entrypoint to the right place
Move the argcomplete entrypoint to where the documentation says it
should be: after the argparse parser has been constructed, but before
the parser has parsed the arguments.

Previously it was called after argparse had parsed the arguments. This
would cause some arguments to not be completed correctly; e.g. "--h"
would no longer be detected for completion when using the upcoming
subprocess completion testing.