Restart subprocess on restart_build
The current implementation of the restart mechanism doesn't rebuild the
action structure. This will lead to unexpected (wrong) results for cases
where early actions request a restart after modifying the tree somehow,
e.g. by fetching and writing a file containing variables.  Global
variables (e.g. list of actions) and other state could then remain from
the previous execution and confuse the restarted build. Any imported
files would also be cached between executions.

Solve this by moving up the restart mechanism so that it restarts the
building of the action on a subprocess level.

The implementation here is restricted in that actions that execute in
the ambient/default environment are no longer allowed to do restarts,
since builds in this environment are not executed in a subprocess.
Implementing this is left for a future expansion if deemed important.

Followup to 578926e121839e766842c5aa482952228eae2f35 and #306.