Rework Python installation in code quality workflow
Using uv for installing Python everywhere interfered with tests creating
venvs themselves, e.g. the environments example. This was because (1) uv
would always be available, making virtualenv never be used there even
though that is one of the test paths in CI, and (2) the system Python
would be used, and often a different version than what we wanted to test
with.

The Python versions available via actions/setup-python lag behind by
quite a lot on macOS, so instead use prototools to install Python where
we don't use uv.

Details:

- Use moonrepo/setup-toolchain@v0 for prototools setup
- Install nox with pip when nox_env_manager is 'virtualenv'
- Upgrade actions/setup-python to v6 for the audit job

Followup to #378 and #383.